Extract shared ListViewController test and Extract common methods to helpers as extension
