+++ EssentialFeed/EssentialFeediOSTests/Feed UI/snapshots/EMPTY_FEED_dark.png
