+++ EssentialFeed/EssentialFeediOSTests/Feed UI/snapshots/EMPTY_FEED_light.png
