+++ EssentialFeed/EssentialFeediOSTests/Shaerd UI/ListSanpshotTests.swift
@@ -1,0 +1,49 @@
+//
+//  ListSanpshotTests.swift
+//  EssentialFeediOSTests
+//
+//  Created by Prabhat Tiwari on 27/09/23.
+//
+
+import XCTest
+import EssentialFeediOS
+@testable import EssentialFeed
+
+final class ListSanpshotTests: XCTestCase {
+
+    func test_emptyFeed() {
+        let sut = makeSUT()
+        
+        sut.display(emptyList())
+        
+        assert(snapshot: sut.snapshot(for: .iPhone8(style: .light)), named: "EMPTY_FEED_light")
+        assert(snapshot: sut.snapshot(for: .iPhone8(style: .dark)), named: "EMPTY_FEED_dark")
+    }
+    
+    func test_feedWithErrorMessage() {
+        let sut = makeSUT()
+        
+        sut.display(.error(message: "This is a\nmulti-line\nerror message"))
+        
+        assert(snapshot: sut.snapshot(for: .iPhone8(style: .light)), named: "LIST_WITH_ERROR_MESSAGE_light")
+        assert(snapshot: sut.snapshot(for: .iPhone8(style: .dark)), named: "LIST_WITH_ERROR_MESSAGE_dark")
+    }
+
+    // MARK: - Helpers
+    
+    func makeSUT(fiel: StaticString = #file, line: UInt = #line) -> ListViewController {
+        
+        let bundel = Bundle(for: ListViewController.self)
+        let storyBoard = UIStoryboard(name: "Feed", bundle: bundel)
+        let controller = storyBoard.instantiateInitialViewController() as! ListViewController
+        controller.loadViewIfNeeded()
+        controller.tableView.showsVerticalScrollIndicator = false
+        controller.tableView.showsHorizontalScrollIndicator = false
+        return controller
+    }
+    
+    private func emptyList() -> [CellController] {
+        return []
+    }
+    
+}
