+++ EssentialFeed/EssentialFeediOSTests/Shaerd UI/snapshots/LIST_WITH_ERROR_MESSAGE_light.png
